requirements.txt
python-dotenv
Pillow

SQLAlchemy
pymysql

git+https://github.com/MaximumMaxxx/Cordless-Quart-Discord 
py-cord==2.2
aiodns
cchardet
brotlipy
orjson

Werkzeug==2.0.0 # Fix quart-csfr
quart==0.17.0
Quart-Csrf